Move enums into enums folder
